[QI-5] Moving view holder to adapter package, not really a view.
