+++ app/src/main/java/it/quip/android/adapter/NotificationBaseViewHolder.java
@@ -1,4 +1,4 @@
-package it.quip.android.view;
+package it.quip.android.adapter;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
